app/requirements.txt
alembic==1.4.3
aniso8601==8.0.0
astroid==2.4.2
certifi==2018.8.24
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
dnspython==2.0.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.9
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
install==1.3.4
isort==5.6.3
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mod-wsgi==4.7.1
numpy==1.18.5
pandas==0.25.3
plaid-python==7.1.0
pycparser==2.20
pylint==2.6.0
pyodbc==4.0.30
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
urllib3==1.25.10
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
